Move the AST operator tables out of `guarded_eval`
`IPython.terminal.shortcuts.filters` evaluates the filter expressions of the
built-in key bindings at import time, and needed three names from
`IPython.core.guarded_eval` to do it: `BINARY_OP_DUNDERS`, `UNARY_OP_DUNDERS`
and `_find_dunder`. Importing `guarded_eval` for them cost ~5ms and pulled in
`typing_extensions` on every terminal start, for two dicts and a six-line
helper that only need `ast`.

Move those (and `COMP_OP_DUNDERS`, which belongs with them) to
`IPython.core._dunder_ops`, and re-export them from `guarded_eval`, which
stays their documented home.
